hw/i386/pc: Let e820_add_entry() return a ssize_t type
e820_add_entry() returns an array size on success, or a negative
value on error.

Message-Id: <20190524063553.5339-4-philmd@redhat.com>

+++ hw/i386/pc.c
@@ -868,7 +868,7 @@ static void handle_a20_line_change(void *opaque, int irq, int level)
     x86_cpu_set_a20(cpu, level);
 }
 
-int e820_add_entry(uint64_t address, uint64_t length, uint32_t type)
+ssize_t e820_add_entry(uint64_t address, uint64_t length, uint32_t type)
 {
     unsigned int index = le32_to_cpu(e820_reserve.count);
     struct e820_entry *entry;
